Moved @background decorator in separate module, added proof of concept intercept class
